feat: Add user filter to list_wallets method
The `list_wallets` method in the `WalletController` now accepts a `Request` object as a parameter, allowing for user filtering. If the requesting user is not a staff member, the `owner_id` parameter is automatically added to the query parameters. This change improves the security and functionality of the wallet listing feature.

See: Modify Wallet to use Handler / Controller / Repository architecture. #61