Add Docker Compose file to load initial data
The Docker Compose file `docker-compose.load_initial_data.yml` is added to the repository. This file includes a service definition for the backend, which runs the necessary commands to migrate the database, load initial data from `initial_data.json`, and start the server. This will simplify the process of setting up the application with initial data.

See: #58